+++ source/extensions/omni.isaac.orbit_tasks/test/wrappers/test_sb3_wrapper.py
@@ -7,13 +7,11 @@ from __future__ import annotations
 
 """Launch Isaac Sim Simulator first."""
 
-import os
 
 from omni.isaac.orbit.app import AppLauncher, run_tests
 
 # launch the simulator
-app_experience = f"{os.environ['EXP_PATH']}/omni.isaac.sim.python.gym.headless.kit"
-app_launcher = AppLauncher(headless=True, experience=app_experience)
+app_launcher = AppLauncher(headless=True)
 simulation_app = app_launcher.app