+++ source/standalone/workflows/rl_games/train.py
@@ -9,9 +9,7 @@ from __future__ import annotations
 
 """Launch Isaac Sim Simulator first."""
 
-
 import argparse
-import os
 
 from omni.isaac.orbit.app import AppLauncher
 
@@ -32,14 +30,8 @@ AppLauncher.add_app_launcher_args(parser)
 # parse the arguments
 args_cli = parser.parse_args()
 
-# launch the simulator
-# load cheaper kit config in headless
-if args_cli.headless:
-    app_experience = f"{os.environ['EXP_PATH']}/omni.isaac.sim.python.gym.headless.kit"
-else:
-    app_experience = f"{os.environ['EXP_PATH']}/omni.isaac.sim.python.kit"
 # launch omniverse app
-app_launcher = AppLauncher(args_cli, experience=app_experience)
+app_launcher = AppLauncher(args_cli)
 simulation_app = app_launcher.app
 
 """Rest everything follows."""